# specs/001-body-recomp-goals/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Body Recomposition Tracking API
  version: 1.0.0
  description: |
    RESTful API for tracking body recomposition goals including cutting and bulking phases.
    
    ## Features
    - User registration and authentication with JWT tokens
    - Multiple body fat calculation methods (Navy, 3-Site, 7-Site Skinfold)
    - Goal creation and tracking (cutting/bulking)
    - Weekly progress logging and visualization
    - Personalized training and diet plans
    
    ## Authentication
    Most endpoints require JWT bearer token authentication. Obtain tokens via `/api/v1/auth/login`.
    
  contact:
    name: Body Recomp API Support
    email: support@bodyrecomp.example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.bodyrecomp.example.com
    description: Production server
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/goals:
    post:
      tags:
        - Goals
      summary: Create a new goal
      description: |
        Create a cutting or bulking goal. The system calculates caloric targets and timeline projections.
        
        **Cutting**: Set target_body_fat_percentage lower than current
        **Bulking**: Set ceiling_body_fat_percentage higher than current
      operationId: createGoal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoalCreate'
            examples:
              cutting:
                summary: Cutting Goal
                value:
                  goal_type: cutting
                  initial_measurement_id: 660e8400-e29b-41d4-a716-446655440001
                  target_body_fat_percentage: 15.0
              bulking:
                summary: Bulking Goal
                value:
                  goal_type: bulking
                  initial_measurement_id: 660e8400-e29b-41d4-a716-446655440001
                  ceiling_body_fat_percentage: 18.0
      responses:
        '201':
          description: Goal created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Goal'
              examples:
                cutting:
                  summary: Cutting Goal Response
                  value:
                    id: 770e8400-e29b-41d4-a716-446655440002
                    user_id: 550e8400-e29b-41d4-a716-446655440000
                    goal_type: cutting
                    status: active
                    initial_measurement_id: 660e8400-e29b-41d4-a716-446655440001
                    initial_body_fat_percentage: 22.5
                    target_body_fat_percentage: 15.0
                    ceiling_body_fat_percentage: null
                    initial_weight_kg: 90.0
                    target_calories: 2200
                    estimated_weeks_to_goal: 40
                    current_body_fat_percentage: 22.5
                    progress_percentage: 0.0
                    weeks_elapsed: 0
                    is_on_track: true
                    started_at: "2025-10-23T08:10:00Z"
                    completed_at: null
                    created_at: "2025-10-23T08:10:00Z"
                    updated_at: "2025-10-23T08:10:00Z"
                bulking:
                  summary: Bulking Goal Response
                  value:
                    id: 770e8400-e29b-41d4-a716-446655440003
                    user_id: 550e8400-e29b-41d4-a716-446655440000
                    goal_type: bulking
                    status: active
                    initial_measurement_id: 660e8400-e29b-41d4-a716-446655440001
                    initial_body_fat_percentage: 12.0
                    target_body_fat_percentage: null
                    ceiling_body_fat_percentage: 18.0
                    initial_weight_kg: 75.0
                    target_calories: 2900
                    estimated_weeks_to_goal: 130
                    current_body_fat_percentage: 12.0
                    progress_percentage: 0.0
                    weeks_elapsed: 0
                    is_on_track: true
                    started_at: "2025-10-23T08:10:00Z"
                    completed_at: null
                    created_at: "2025-10-23T08:10:00Z"
                    updated_at: "2025-10-23T08:10:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: User already has an active goal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                detail: You already have an active goal. Complete or abandon it before creating a new one.
                error_code: ACTIVE_GOAL_EXISTS
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/goals/{goal_id}:
    get:
      tags:
        - Goals
      summary: Get goal details
      description: Retrieve detailed information about a specific goal including current progress
      operationId: getGoal
      parameters:
        - name: goal_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Goal unique identifier
          example: 770e8400-e29b-41d4-a716-446655440002
      responses:
        '200':
          description: Goal retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Goal'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT authentication token obtained from /api/v1/auth/login
  schemas:
...
    GoalCreate:
      type: object
      required:
        - goal_type
        - initial_measurement_id
      properties:
        goal_type:
          $ref: '#/components/schemas/GoalType'
        initial_measurement_id:
          type: string
          format: uuid
          description: ID of the initial body measurement
          example: 660e8400-e29b-41d4-a716-446655440001
        target_body_fat_percentage:
          type: number
          format: float
          minimum: 3
          maximum: 50
          nullable: true
          description: Target body fat % (required for cutting, must be < current BF%)
          example: 15.0
        ceiling_body_fat_percentage:
          type: number
          format: float
          minimum: 3
          maximum: 50
          nullable: true
          description: Maximum body fat % ceiling (required for bulking, must be > current BF%)
          example: 18.0
      description: |
        For cutting goals: provide target_body_fat_percentage (must be lower than initial)
        For bulking goals: provide ceiling_body_fat_percentage (must be higher than initial)
    Goal:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 770e8400-e29b-41d4-a716-446655440002
        user_id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        goal_type:
          $ref: '#/components/schemas/GoalType'
        status:
          $ref: '#/components/schemas/GoalStatus'
        initial_measurement_id:
          type: string
          format: uuid
          example: 660e8400-e29b-41d4-a716-446655440001
        initial_body_fat_percentage:
          type: number
          format: float
          description: Starting body fat percentage
          example: 22.5
        target_body_fat_percentage:
          type: number
          format: float
          nullable: true
          description: Target body fat (cutting only)
          example: 15.0
        ceiling_body_fat_percentage:
          type: number
          format: float
          nullable: true
          description: Maximum body fat ceiling (bulking only)
          example: null
        initial_weight_kg:
          type: number
          format: float
          description: Starting weight in kilograms
          example: 90.0
        target_calories:
          type: integer
          minimum: 1200
          maximum: 5000
          description: Daily calorie target (deficit for cutting, surplus for bulking)
          example: 2200
        estimated_weeks_to_goal:
          type: integer
          nullable: true
          description: Estimated weeks to reach goal
          example: 40
        current_body_fat_percentage:
          type: number
          format: float
          description: Current body fat percentage (computed field)
          example: 22.5
        progress_percentage:
          type: number
          format: float
          description: Progress toward goal as percentage (computed field)
          example: 0.0
        weeks_elapsed:
          type: integer
          description: Number of weeks since goal started (computed field)
          example: 0
        is_on_track:
          type: boolean
          description: Whether progress is on track (computed field)
          example: true
        started_at:
          type: string
          format: date-time
          example: "2025-10-23T08:10:00Z"
        completed_at:
          type: string
          format: date-time
          nullable: true
          example: null
        created_at:
          type: string
          format: date-time
          example: "2025-10-23T08:10:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2025-10-23T08:10:00Z"
...
    GoalType:
      type: string
      enum:
        - cutting
        - bulking
      description: |
        Type of body recomposition goal:
        - cutting: Fat loss phase (caloric deficit)
        - bulking: Muscle gain phase (caloric surplus)
      example: cutting
    GoalStatus:
      type: string
      enum:
        - active
        - completed
        - abandoned
      description: |
        Goal lifecycle status:
        - active: Currently working toward goal
        - completed: Successfully reached target/ceiling
        - abandoned: Goal discontinued before completion
      example: active
    Error:
      type: object
      properties:
        detail:
          type: string
          description: Human-readable error message
          example: The requested resource was not found
        error_code:
          type: string
          description: Machine-readable error code
          example: RESOURCE_NOT_FOUND
        field:
          type: string
          nullable: true
          description: Field name if validation error
          example: target_body_fat_percentage
        value:
          nullable: true
          description: Invalid value if applicable
          example: 3.5
    ValidationError:
      type: object
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
                items:
                  type: string
                description: Location of the validation error
                example: ["body", "height_cm"]
              msg:
                type: string
                description: Validation error message
                example: ensure this value is greater than or equal to 120
              type:
                type: string
                description: Error type
                example: value_error.number.not_ge
              ctx:
                type: object
                nullable: true
                description: Additional context about the error
  responses:
    BadRequest:
      description: Bad request - Invalid input or business logic violation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            unsafe_target:
              summary: Unsafe body fat target
              value:
                detail: "Target body fat percentage is below safe threshold for males (8%)"
                error_code: UNSAFE_BODY_FAT_TARGET
                field: target_body_fat_percentage
                value: 5.0
            missing_measurements:
              summary: Missing required measurements
              value:
                detail: "Navy method requires waist_cm and neck_cm measurements"
                error_code: MISSING_REQUIRED_MEASUREMENTS
                field: null
                value: null
    Unauthorized:
      description: Authentication required or token invalid/expired
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            missing_token:
              summary: Missing authentication token
              value:
                detail: "Authentication token is required"
                error_code: AUTHENTICATION_REQUIRED
            invalid_token:
              summary: Invalid or expired token
              value:
                detail: "Authentication token is invalid or expired"
                error_code: INVALID_TOKEN
            invalid_credentials:
              summary: Invalid login credentials
              value:
                detail: "Incorrect email or password"
                error_code: INVALID_CREDENTIALS
    Forbidden:
      description: Insufficient permissions to access the resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "You do not have permission to access this resource"
            error_code: PERMISSION_DENIED
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            goal_not_found:
              summary: Goal not found
              value:
                detail: "Goal with ID 770e8400-e29b-41d4-a716-446655440002 not found"
                error_code: GOAL_NOT_FOUND
            measurement_not_found:
              summary: Measurement not found
              value:
                detail: "Measurement with ID 660e8400-e29b-41d4-a716-446655440001 not found"
                error_code: MEASUREMENT_NOT_FOUND
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
          example:
            detail:
              - loc: ["body", "height_cm"]
                msg: "ensure this value is greater than or equal to 120"
                type: "value_error.number.not_ge"
                ctx:
                  limit_value: 120
              - loc: ["body", "email"]
                msg: "value is not a valid email address"
                type: "value_error.email"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "An unexpected error occurred. Please try again later."
            error_code: INTERNAL_SERVER_ERROR